requirements.txt
flask
flask-cors
requests
schedule
python-dotenv
groq
gunicorn
